Refactor rds2_stream.py into modular package with improved CLI and imports
